Move colorspace converters to colorspaces.py
